Create stormbreaker package, refactor Flask app
Refactor the Flask app into a Python package, and move project resources into the `/stormbreaker` directory.
