+++ stormbreaker/.gitignore
@@ -1,3 +1,4 @@
+.venv/
 venv/
 
 *.pyc
