+++ stormbreaker/stormbreaker/__init__.py
@@ -1,0 +1,10 @@
+"""Stormbreaker App."""
+from flask import Flask
+
+
+app = Flask(__name__)
+
+
+# Import Views
+import stormbreaker.views.api.cloud_accounts    # noqa
+import stormbreaker.views.api.dcloud_sessions   # noqa
