+++ stormbreaker/stormbreaker/static/swagger.yaml
@@ -5,7 +5,7 @@ info:
   version: 0.1.0
 
 
-host: fixme.cisco.com
+host: stormbreaker.cisco.com
 basePath: /api
 schemes:
   - http
